Move tensors and helpers to math module
