+++ apsg/math/vector.py
@@ -3,20 +3,17 @@
 
 """
 A vector algebra types and functions.
-
-== Contains
-
-- ``Vector``
-
-- ``Vector2``
-
-- ``Vector3``
-
 """
 
 
 from apsg.math.scalar import Scalar
 from apsg.math.matrix import Matrix
+
+
+class NonConformableVectors(Exception):
+    """
+    Raises when vectors are not conformable for a certain operation.
+    """
 
 
 class Vector(Matrix):
@@ -65,10 +62,18 @@ class Vector2(Vector):
     def y(self):
         return self[1]
 
+    @classmethod
+    def unit_x(cls):
+        return cls(1, 0)
+
+    @classmethos
+    def unit_y(cls):
+        return cls(0, 1)
+
 
 class Vector3(Vector):
 
-    __shape__ = (2, 1)
+    __shape__ = (3, 1)
 
     __slots__ = ("_elements",)
 
@@ -80,7 +85,7 @@ class Vector3(Vector):
 
     def __pow__(self, other): # (Vector3) -> Vector3
         """
-        Calculate the vector product.
+        Calculate the vector product between ``self`` and ``other`` vector.
 
         Returns:
             The vector product of ``self`` and ``other`` vector.
@@ -91,14 +96,17 @@ class Vector3(Vector):
             self.x * other.y - self.y * other.x
         )
 
-    def cross(self, other): # (Vector3) -> Vector3
-        """
-         Calculate the vector product.
+    @classmethod
+    def unit_x(cls): # e1
+        return cls(1, 0, 0)
 
-        Returns:
-            The vector product of ``self`` and ``other`` vector.
-        """
-        return self ** other
+    @classmethos
+    def unit_y(cls): # e2
+        return cls(0, 1, 0)
+
+     @classmethod
+    def unit_z(cls): # e3
+        return cls(0, 0, 1)
 
     @property
     def x(self):
@@ -110,13 +118,42 @@ class Vector3(Vector):
 
     @property
     def z(self):
+        return self[2]
+
+    def cross(self, other): # (Vector3) -> Vector3
+        """
+        Calculate the vector product between ``self`` and ``other`` vector.
+
+        Returns:
+            The vector product of ``self`` and ``other`` vector.
+        """
+        return self ** other
+
+
+class Vector4(Vector):
+
+    __shape__ = (4, 1)
+
+    __slots__ = ("_elements",)
+
+    def __init__(self, *elements):
+        if len(elements) > 3:
+            raise Exception("Wrong number of elements")
+
+        super(Vector3, self).__init__(*elements)
+
+    @property
+    def x(self):
+        return self[0]
+
+    @property
+    def y(self):
         return self[1]
 
+    @property
+    def z(self):
+        return self[2]
 
-if __name__ == '__main__':
-    v1 = Vector3(1, 0, 0)
-    v2 = Vector3(0, 1, 0)
-
-    print(v1 % v2)
-
-    v1 = Vector3(1, 2, 3)
+    @property
+    def w(self):
+        return self[3]
